# feat: optimize common case of GlobPath