- [x] Have you signed the [CLA](http://www.ubuntu.com/legal/contributors/)?

-----
Results of the benchmark on my machine (using hyperfine with 10 rounds):
```
$ bash benchmark.sh HEAD main 
Creating rev: e2ee603c7396b33038e47352c0722b5b1202fbfe (main)
Creating rev: 19829794e6454f78334c9a67b74d8abb9bc66b25 (HEAD)
Benchmark 1: ./19829794e6454f78334c9a67b74d8abb9bc66b25 info --release ../chisel-releases/ubuntu-24.04 'python3.12_core'
  Time (mean ± σ):      5.341 s ±  0.115 s    [User: 5.528 s, System: 0.028 s]
  Range (min … max):    5.239 s …  5.635 s    10 runs
 
Benchmark 2: ./e2ee603c7396b33038e47352c0722b5b1202fbfe info --release ../chisel-releases/ubuntu-24.04 'python3.12_core'
  Time (mean ± σ):     18.366 s ±  0.139 s    [User: 19.677 s, System: 0.103 s]
  Range (min … max):   18.219 s … 18.675 s    10 runs
 
Summary
  ./19829794e6454f78334c9a67b74d8abb9bc66b25 info --release ../chisel-releases/ubuntu-24.04 'python3.12_core' ran
    3.44 ± 0.08 times faster than ./e2ee603c7396b33038e47352c0722b5b1202fbfe info --release ../chisel-releases/ubuntu-24.04 'python3.12_core'
```
Basically, **HEAD takes 5.341s on my machine compared to 18.366s for main**.